fix some bugs

+++ common/utils/database/pg/test/set1/set1.go
@@ -44,33 +44,50 @@ func CreateUserSampleData(db *gorm.DB) ([]utilDB.User, error) {
 }
 
 func ACLUsersTable(t *testing.T, cfg *backendConfig.Config) func() {
-	db := cfg.DB(utilDB.DBName)
-
 	var (
-		schemaName = cfg.DBSchemaName
-		username   = cfg.DBUsername
-		password   = cfg.DBPassword
+		isCreateUser = false
+		username     = cfg.DBUsername
+		password     = cfg.DBPassword
 	)
+	for _, dbase := range cfg.Databases {
+		db := cfg.DB(dbase.DBName)
+		schemaName := dbase.SchemaName
 
-	if err := db.Exec(fmt.Sprintf("CREATE ROLE %s LOGIN PASSWORD '%s';", username, password)).Error; err != nil {
-		t.Fatalf("Fail to create ROLE. %s", err.Error())
+		if !isCreateUser {
+			if err := db.Exec(fmt.Sprintf("CREATE ROLE %s LOGIN PASSWORD '%s';", username, password)).Error; err != nil {
+				t.Fatalf("Fail to create ROLE. %s", err.Error())
+			}
+			isCreateUser = true
+		}
+
+		if err := db.Exec(fmt.Sprintf("GRANT USAGE ON SCHEMA %s TO %s ;", schemaName, username)).Error; err != nil {
+			t.Fatalf("Fail to create ROLE. %s", err.Error())
+		}
+		if err := db.Exec(fmt.Sprintf("GRANT USAGE, SELECT ON ALL SEQUENCES IN SCHEMA %s TO %s;", schemaName, username)).Error; err != nil {
+			t.Fatalf("Fail to GRANT SCHEMA. %s", err.Error())
+		}
+		if err := db.Exec(fmt.Sprintf("GRANT SELECT, INSERT ON users TO %s", username)).Error; err != nil {
+			t.Fatalf("Fail to GRANT SELECT, CREATE ON users. %s", err.Error())
+		}
 	}
-	if err := db.Exec(fmt.Sprintf("GRANT USAGE ON SCHEMA %s TO %s ;", schemaName, username)).Error; err != nil {
-		t.Fatalf("Fail to create ROLE. %s", err.Error())
-	}
-	if err := db.Exec(fmt.Sprintf("GRANT USAGE, SELECT ON ALL SEQUENCES IN SCHEMA %s TO %s;", schemaName, username)).Error; err != nil {
-		t.Fatalf("Fail to GRANT SCHEMA. %s", err.Error())
-	}
-	if err := db.Exec(fmt.Sprintf("GRANT SELECT, INSERT ON users TO %s", username)).Error; err != nil {
-		t.Fatalf("Fail to GRANT SELECT, CREATE ON users. %s", err.Error())
-	}
+
 	return func() {
-		db.Exec(fmt.Sprintf("REASSIGN OWNED BY %s TO postgres;", username))
-		db.Exec(fmt.Sprintf("DROP OWNED BY %s;", username))
-
-		err := db.Exec(fmt.Sprintf("DROP ROLE IF EXISTS %s;", username)).Error
+		cfg.DBUsername = "postgres"
+		cfg.DBPassword = "example"
+		err := cfg.UpdateDB()
 		if err != nil {
-			t.Fatalf("Fail to drop ROLE. %s", err.Error())
+			t.Fatalf("Fail to update connection. %s", err.Error())
+		}
+		for _, dbase := range cfg.Databases {
+			db := cfg.DB(dbase.DBName)
+			db.Exec(fmt.Sprintf("REASSIGN OWNED BY %s TO postgres;", username))
+			db.Exec(fmt.Sprintf("DROP OWNED BY %s;", username))
+		}
+		if len(cfg.Databases) > 0 {
+			err := cfg.DB(cfg.Databases[0].DBName).Exec(fmt.Sprintf("DROP ROLE IF EXISTS %s;", username)).Error
+			if err != nil {
+				t.Fatalf("Fail to drop ROLE. %s", err.Error())
+			}
 		}
 	}
 }
@@ -138,7 +155,16 @@ func CreateConfig(t *testing.T) (*backendConfig.Config, func()) {
 					Role:     "admin",
 					DatabaseList: []backendConfig.Database{
 						{
-							DBName: "test",
+							DBName: "test1",
+							Tables: []backendConfig.Table{
+								{
+									TableName: "users",
+									ACL:       "cru",
+								},
+							},
+						},
+						{
+							DBName: "test2",
 							Tables: []backendConfig.Table{
 								{
 									TableName: "users",
